--- a/cmd/cronosd/opendb/opendb.go
+++ b/cmd/cronosd/opendb/opendb.go
@@ -6,15 +6,17 @@ package opendb
 import (
 	"path/filepath"
 
+	"github.com/cosmos/cosmos-sdk/server/types"
+	"github.com/cosmos/cosmos-sdk/simapp"
 	dbm "github.com/tendermint/tm-db"
 )
 
-func OpenDB(home string, backendType dbm.BackendType) (dbm.DB, error) {
+func OpenDB(opts types.AppOptions, home string, backendType dbm.BackendType) (dbm.DB, error) {
 	dataDir := filepath.Join(home, "data")
 	return dbm.NewDB("application", backendType, dataDir)
 }
 
 // OpenReadOnlyDB opens rocksdb backend in read-only mode.
 func OpenReadOnlyDB(home string, backendType dbm.BackendType) (dbm.DB, error) {
-	return OpenDB(home, backendType)
+	return OpenDB(simapp.EmptyAppOptions{}, home, backendType)
 }