--- a/x/cronos/keeper/grpc_query.go
+++ b/x/cronos/keeper/grpc_query.go
@@ -11,6 +11,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/crypto-org-chain/cronos/v2/x/cronos/types"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/evmos/ethermint/x/evm/keeper"
 	evmtypes "github.com/evmos/ethermint/x/evm/types"
 )
 
@@ -65,7 +66,6 @@ func (k Keeper) ReplayBlock(goCtx context.Context, req *types.ReplayBlockRequest
 	blockHeight := big.NewInt(req.BlockNumber)
 	homestead := ethCfg.IsHomestead(blockHeight)
 	istanbul := ethCfg.IsIstanbul(blockHeight)
-	london := ethCfg.IsLondon(blockHeight)
 	evmDenom := params.EvmDenom
 
 	// we assume the message executions are successful, they are filtered in json-rpc api
@@ -81,15 +81,12 @@ func (k Keeper) ReplayBlock(goCtx context.Context, req *types.ReplayBlockRequest
 			return nil, err
 		}
 
-		if _, _, err := k.evmKeeper.DeductTxCostsFromUserBalance(
-			ctx,
-			*msg,
-			txData,
-			evmDenom,
-			homestead,
-			istanbul,
-			london,
-		); err != nil {
+		fees, err := keeper.VerifyFee(txData, evmDenom, nil, homestead, istanbul, ctx.IsCheckTx())
+		if err != nil {
+			return nil, err
+		}
+
+		if err := k.evmKeeper.DeductTxCostsFromUserBalance(ctx, fees, common.HexToAddress(msg.From)); err != nil {
 			return nil, err
 		}