--- a/x/cronos/keeper/ibc.go
+++ b/x/cronos/keeper/ibc.go
@@ -11,8 +11,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	errorsmod "cosmossdk.io/errors"
-	ibctransfertypes "github.com/cosmos/ibc-go/v5/modules/apps/transfer/types"
-	ibcclienttypes "github.com/cosmos/ibc-go/v5/modules/core/02-client/types"
+	ibctransfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
+	ibcclienttypes "github.com/cosmos/ibc-go/v6/modules/core/02-client/types"
 	"github.com/crypto-org-chain/cronos/v2/x/cronos/types"
 )
 
@@ -179,13 +179,20 @@ func (k Keeper) ibcSendTransfer(ctx sdk.Context, sender sdk.AccAddress, destinat
 	params := k.GetParams(ctx)
 	timeoutTimestamp := uint64(ctx.BlockTime().UnixNano()) + params.IbcTimeout
 	timeoutHeight := ibcclienttypes.ZeroHeight()
-	return k.transferKeeper.SendTransfer(
-		ctx,
-		ibctransfertypes.PortID,
-		channelID,
-		coin,
-		sender,
-		destination,
-		timeoutHeight,
-		timeoutTimestamp)
+	if _, err = k.transferKeeper.Transfer(
+		sdk.WrapSDKContext(ctx),
+		&ibctransfertypes.MsgTransfer{
+			SourcePort:       ibctransfertypes.PortID,
+			SourceChannel:    channelID,
+			Token:            coin,
+			Sender:           sender.String(),
+			Receiver:         destination,
+			TimeoutHeight:    timeoutHeight,
+			TimeoutTimestamp: timeoutTimestamp,
+		},
+	); err != nil {
+		return err
+	}
+
+	return nil
 }