--- a/x/cronos/keeper/mock/ibckeeper_mock.go
+++ b/x/cronos/keeper/mock/ibckeeper_mock.go
@@ -2,8 +2,8 @@ package keeper_test
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/ibc-go/v5/modules/apps/transfer/types"
-	clienttypes "github.com/cosmos/ibc-go/v5/modules/core/02-client/types"
+	"github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
+	clienttypes "github.com/cosmos/ibc-go/v6/modules/core/02-client/types"
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 )