--- a/x/cronos/keeper/params.go
+++ b/x/cronos/keeper/params.go
@@ -5,7 +5,7 @@ import (
 
 	"cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	transferTypes "github.com/cosmos/ibc-go/v5/modules/apps/transfer/types"
+	transferTypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
 	"github.com/crypto-org-chain/cronos/v2/x/cronos/types"
 	evmTypes "github.com/evmos/ethermint/x/evm/types"
 )