@cryptocj 2023-08-12
## PR Analysis

- 🎯 **Main theme:** Benchmarking optimization for sender cache on ethermint main branch
- 📌 **Type of PR:** Enhancement
- ✨ **Focused PR:** True
- 🔒 **Security concerns:** No security concerns found
## PR Feedback

-  **General suggestions:**

  - It would be helpful to provide more context in the PR description about the benchmark results and the optimization being implemented.
  - Consider adding relevant tests to ensure the correctness and performance of the optimization.

- 🤖 **Code feedback:**


  - **relevant file:** ``app/app.go``
   **suggestion:** Consider adding a comment to explain the purpose of the `RandomGenesisAccounts` function and how it is used in the code. [medium]
   **relevant line:** // don't need that much entropy for simulation


  - **relevant file:** ``app/sim_test.go``
   **suggestion:** Update the comment in the `RandomAccounts` function to explain the purpose of generating random accounts with an ethsecp256k1 private key. [medium]
   **relevant line:** [// TODO: replace secp256k1.GenPrivKeyFromSecret() with similar function in go-ethereum](https://github.com/crypto-org-chain/cronos/pull/933/files#diff-05389192d5e22a5bedc5409d8f7647c4149c6d5f65bac9ba91fba0fc75df0be6R408)


  - **relevant file:** ``x/cronos/keeper/ibc.go``
   **suggestion:** Consider adding error handling for the `Transfer` function call to handle any potential errors that may occur during the transfer. [medium]
   **relevant line:** [if _, err = k.transferKeeper.Transfer(](https://github.com/crypto-org-chain/cronos/pull/933/files#diff-8dafea39e7d85f7fcf0a92f288c2004a7a678e04de5e2afdf5a5929179bacf73R182)